acceptance framework improvements
This commit adds acceptance framework improvements needed to
add acceptance tests for the following PR, which will add
--globs as an argument to the extract-translations command.

- removes shell-quote, because it was parsing arguments like
  **/myglob.hbs as a special glob object, when all we need is a string
- adds globalSetup/globalTeardown to jest, which will do one time setup
  for the test-themes. Trying to do it in runPlayground() caused issues
  where different jest test instances would try to perform git operations
  on the same test theme and clobber each other. This is preferable to
  running tests serially with --runInBand
- sets jambo's log level to 'silent' for acceptance tests, because its noisy
  by default

J=SLAP-1378
TEST=auto

reran tests
