+++ package.json
@@ -46,14 +46,15 @@
     "@types/jest": "^26.0.23",
     "eslint": "^7.8.1",
     "jest": "^25.4.0",
-    "prettier": "^2.2.1",
-    "shell-quote": "^1.7.2"
+    "prettier": "^2.2.1"
   },
   "jest": {
     "collectCoverageFrom": [
       "src/**"
     ],
     "verbose": true,
+    "globalSetup": "./tests/setup/globalsetup.js",
+    "globalTeardown": "./tests/setup/globalteardown.js",
     "setupFilesAfterEnv": [
       "./tests/setup/setup.js"
     ],
