+++ tests/acceptance/setup/TestInstance.js
@@ -1,5 +1,4 @@
 const buildJamboCLI = require('../../../src/buildJamboCLI');
-const { parse } = require('shell-quote');
 const { error } = require('../../../src/utils/logger');
 
 /**
@@ -8,9 +7,8 @@ const { error } = require('../../../src/utils/logger');
  */
 module.exports = class TestInstance {
   async jambo(command) {
-    const commandArgs = parse(command);
+    const commandArgs = command.split(/\s/);
     const argv = [process.argv[0], process.argv[1], ...commandArgs];
-
     return buildJamboCLI(argv)
       .scriptName('jambo')
       .fail(function(msg, err) {
