+++ tests/acceptance/suites/basic-flow.js
@@ -2,9 +2,6 @@ const fs = require('fs');
 
 const { runInPlayground } = require('../setup/playground');
 
-// silence jambo's noisy output
-console.log = jest.fn();
-
 it('can init -> import -> create page -> build', () => runInPlayground(async t => {
   await t.jambo('init');
   await t.jambo('import --themeUrl ../test-themes/basic-flow');
