+++ tests/acceptance/suites/describe.js
@@ -8,7 +8,8 @@ it('describe on default and custom commands', () => runInPlayground(async t => {
   await t.jambo('init');
   await t.jambo('import --themeUrl ../test-themes/describe');
   await t.jambo('describe');
+  const receivedJSON = JSON.parse(console.log.mock.calls[0][0]);
   const expectedJSON = 
-    fs.readFileSync('../fixtures/describe/describe-test.json', 'utf-8');
-  expect(console.log).toHaveBeenCalledWith(expectedJSON);
+    JSON.parse(fs.readFileSync('../fixtures/describe/describe-test.json', 'utf-8'));
+  expect(receivedJSON).toEqual(expectedJSON);
 }));
