+++ tests/setup/globalsetup.js
@@ -1,0 +1,5 @@
+const { setupTestThemes } = require('./themesetup');
+
+module.exports = async () => {
+  await setupTestThemes();
+}
