+++ tests/setup/globalteardown.js
@@ -1,0 +1,6 @@
+const { cleanupTestThemes } = require('./themesetup');
+
+module.exports = async () => {
+  cleanupTestThemes();
+}
+
